refactored: move task-card component from components to partials on boards page
